refactor into modules, cleanup names
