+++ src/py3r/behaviour/features/features.py
@@ -2297,8 +2297,8 @@ class Features:
 
         ```
         """
-        from py3r.behaviour.animation.geometry_stream import (
-            build_geometry_stream_from_points,
+        from py3r.behaviour.animation import (
+            build_animation_stream_from_points,
             collect_dynamic_source_names_from_style,
         )
 
@@ -2344,7 +2344,7 @@ class Features:
                         raise ValueError(
                             f"Dynamic style source '{name}' not found in Features or Tracking data"
                         )
-        return build_geometry_stream_from_points(
+        return build_animation_stream_from_points(
             points=points_arr,
             point_names=point_names,
             draw_points=points,
