+++ tests/test_animation_stream.py
@@ -3,10 +3,10 @@ import types
 import numpy as np
 import pandas as pd
 
-from py3r.behaviour.animation.geometry_stream import (
+from py3r.behaviour.animation.animation_stream import (
     _format_overlay_value,
-    build_geometry_stream,
-    build_geometry_stream_from_points,
+    build_animation_stream,
+    build_animation_stream_from_points,
 )
 from py3r.behaviour.features.features import Features
 from py3r.behaviour.tracking.tracking import Tracking
@@ -86,7 +86,7 @@ def test_geometry_animation_stream_read_reset_iter():
         ],
         dtype=float,
     )
-    stream = build_geometry_stream_from_points(
+    stream = build_animation_stream_from_points(
         points=points,
         point_names=["nose", "tail"],
         draw_points=["nose"],
@@ -110,7 +110,7 @@ def test_geometry_animation_stream_read_reset_iter():
 
 def test_geometry_animation_stream_render_into_copy_modes():
     points = np.array([[[10.0, 10.0]]], dtype=float)
-    stream = build_geometry_stream_from_points(
+    stream = build_animation_stream_from_points(
         points=points,
         point_names=["nose"],
         draw_points=["nose"],
@@ -150,7 +150,7 @@ def test_tracking_points_to_numpy_undo_meta_scaling_inverts_meta_scaling():
     assert float(out[0, 0, 2]) == 9.0
 
 
-def test_build_geometry_stream_wrapper_supports_3d_points():
+def test_build_animation_stream_wrapper_supports_3d_points():
     df = pd.DataFrame(
         {
             "p1.x": [0.0, 1.0],
@@ -162,7 +162,7 @@ def test_build_geometry_stream_wrapper_supports_3d_points():
         },
         index=pd.Index([10, 11], name="frame"),
     )
-    stream = build_geometry_stream(
+    stream = build_animation_stream(
         df,
         point_names=["p1"],
         lines=[("p1", "p2")],
@@ -293,7 +293,7 @@ def test_dynamic_boundary_style_from_feature_source():
     points = np.empty((3, 0, 2), dtype=float)
     poly = np.array([[10.0, 10.0], [50.0, 10.0], [50.0, 40.0], [10.0, 40.0]], dtype=float)
     boundary_arrays = [("zone", np.repeat(poly[None, :, :], 3, axis=0))]
-    stream = build_geometry_stream_from_points(
+    stream = build_animation_stream_from_points(
         points=points,
         point_names=[],
         frame_ids=np.array([0, 1, 2]),
